multichat_system/requirements.txt
Django>=4.2,<5.0
mysqlclient>=2.2.0
python-decouple
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-filter
requests 